[update] createartgram 들어가기 전에, 아트그램 및 메인페이지 삭제파일 정리 106
