+++ README.md
@@ -16,23 +16,20 @@
 - ✅전시회 상세 페이지에서 좋아요, 스크랩 (완료.)
 
 ## 박영찬: 아트그램 페이지, 메인페이지
-
 ### 아트그램
-
-- 아트그램 페이지 기초 레이아웃 생성
-- 아트그램 POST 기능구현 : Drag & Drop, S3bucket upload, refactoring
-- 아트그램 GET 기능구현, dayjs로 시간제어 -> (디자인 적용이후 수정/삭제 기능구현 예정)
-- 아트그램 상세모달페이지 기초 레이아웃 구현
-- 아트그램 상세모달페이지 댓글 POST 기능구현
-- 아트그램 상세모달페이지 댓글 GET 기능구현
-- 아트그램 페이지, 상세모달페이지 GET 방식에 있어서 GET 최적화 방식으로 세부설정(필요한 곳에서만 요청이 이뤄지도록)
-- 아트그램 상세모달페이지, 복수의 이미지에 대해서 슬라이더 구현
-- 아트그램 상세모달페이지 댓글 PATCH - Delete, Update 기능구현
+- 아트그램, 상세모달페이지, 아트그램 생성페이지 기초 레이아웃 생성
+- 아트그램 GET,POST,DELETE 기능구현(UPDATE는 디자인 시안이 확정되지 않았음)
+- 아트그램 댓글 GET,POST,DELETE,UPDATE 기능구현
+- 아트그램 대댓글 GET,POST,DELETE 기능구현(UPDATE는 적용할지 논의 중)
+- 아트그램, 댓글, 대댓글이 각각의 상황에서만 각각의 요소를 서버에 요청하도록 기능과 로직 분리 
+- 아트그램 상세모달페이지, 복수의 이미지에 대해서 슬라이더 적용
 - (2-3주차) 아트그램, 상세모달페이지 디자인 시안 변경에 따른 디자인 적용
+- (3주차) 아트그램 등록 페이지 디자인 시안에 따른 디자인 적용 
 
 ### 메인페이지
-
-- 메인페이지
+- 메인페이지, 5개의 article에 대한 각각의 GET 요청 기능구현
+- 메인페이지, 5개의 article에 대한 각각의 react-slick 적용
+- 메인페이지, 아트그램 상세모달페이지 연결 및, 기능구현 연결 확인
 
 ## 김재란: 로그인,마이 페이지
