--- src/features/artgram/ArtgramInp.jsx
@@ -1,24 +1,0 @@
-import React from 'react'
-
-function ArtgramInp({selectFile, onChange}) {
-    const imgBtnhandle = () => {
-      selectFile.current.click();
-  };
-  return (
-    <>
-      <input
-        style={{ display: "none" }}
-        ref={selectFile}
-        type="file"
-        onChange={onChange}
-        multiple
-        accept="image/*"
-      />
-      <button type="button" onClick={imgBtnhandle}>
-        사진등록
-      </button>
-    </>
-  );
-}
-
-export default ArtgramInp
