--- src/features/artgram/ArtgramShowimg.jsx
@@ -1,19 +1,0 @@
-import React from 'react'
-
-function ArtgramShowimg({img, index, onCilck}) {
-  return (
-    <div
-      key={`image${index}`}
-      style={{ position: "relative", margin: "20px 0" }}
-    >
-      <img src={img} alt="업로드이미지" width="235px" />
-      <div style={{ position: "absolute", top: "0", right: "0" }}>
-        <button onClick={onCilck} data-index={index}>
-          삭제
-        </button>
-      </div>
-    </div>
-  );
-}
-
-export default ArtgramShowimg
