--- src/features/main/mainpageexample/fifithSliderList.js
@@ -1,8 +1,0 @@
-export const fifithSliderList = [
-  {artgramId:1, title:"첫째", authorNickName : "Edwin-Author", imgUrl: "https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png"},
-  {artgramId:2, title:"둘째", authorNickName : "Edwin-Author", imgUrl: "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200"},
-  {artgramId:3, title:"셋째", authorNickName : "Edwin-Author", imgUrl: "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619"},
-  {artgramId:4, title:"넷째", authorNickName : "Edwin-Author", imgUrl: "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200"},
-  {artgramId:5, title:"다섯째", authorNickName : "Edwin-Author", imgUrl: "https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png"},
-  {artgramId:6, title:"여섯째", authorNickName : "Edwin-Author", imgUrl: "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619"}
-]
