--- src/features/main/mainpageexample/firstSliderList.js
@@ -1,57 +1,0 @@
-export const firstSliderList = [
-  {
-    num:1,
-    exhibitionTitle: "전시회 제목1",
-    exhibitionDesc: "1111 ⟪대지의 시간⟫은 기후변화와 펜데믹 등 전 지구적 위기 시대를 맞이하여 새로운 시대정신으로 떠오르고 있는 '생태학적 세계관'을 탐색하는 장으로서, '공생', '연결', '균형의 회복'을 지향하는 국내외 작가 16명의 작품과 아카이브를 선보인다.",
-    exhibitionDetaillink : "/exhibition/:id",
-    startDate: "2023-04-12T00:00:00.000Z",
-    endDate: "2023-04-13T00:00:00.000Z",
-    address: "경기 성남시 수정구 여수대로 4",
-    writerNickName:"Edwin",
-    imglink:"https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619"
-    },
-  {
-    num:2,
-    exhibitionTitle: "전시회 제목2",
-    exhibitionDesc: "2222 ⟪대지의 시간⟫은 기후변화와 펜데믹 등 전 지구적 위기 시대를 맞이하여 새로운 시대정신으로 떠오르고 있는 '생태학적 세계관'을 탐색하는 장으로서, '공생', '연결', '균형의 회복'을 지향하는 국내외 작가 16명의 작품과 아카이브를 선보인다.",
-    exhibitionDetaillink : "/exhibition/:id",
-    startDate: "2023-04-12T00:00:00.000Z",
-    endDate: "2023-04-13T00:00:00.000Z",
-    address: "경기 성남시 수정구 여수대로 4",
-    writerNickName:"Edwin",
-    imglink:"https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200"
-  },
-  {
-      num:3,
-      exhibitionTitle: "전시회 제목3",
-      exhibitionDesc: "3333 ⟪대지의 시간⟫은 기후변화와 펜데믹 등 전 지구적 위기 시대를 맞이하여 새로운 시대정신으로 떠오르고 있는 '생태학적 세계관'을 탐색하는 장으로서, '공생', '연결', '균형의 회복'을 지향하는 국내외 작가 16명의 작품과 아카이브를 선보인다.",
-      exhibitionDetaillink : "/exhibition/:id",
-      startDate: "2023-04-12T00:00:00.000Z",
-      endDate: "2023-04-13T00:00:00.000Z",
-      address: "경기 성남시 수정구 여수대로 4",
-      writerNickName:"Edwin",
-      imglink:"https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png"
-  },
-  {
-    num:4,
-    exhibitionTitle: "전시회 제목4",
-    exhibitionDesc: "4444 ⟪대지의 시간⟫은 기후변화와 펜데믹 등 전 지구적 위기 시대를 맞이하여 새로운 시대정신으로 떠오르고 있는 '생태학적 세계관'을 탐색하는 장으로서, '공생', '연결', '균형의 회복'을 지향하는 국내외 작가 16명의 작품과 아카이브를 선보인다.",
-    exhibitionDetaillink : "/exhibition/:id",
-    startDate: "2023-04-12T00:00:00.000Z",
-    endDate: "2023-04-13T00:00:00.000Z",
-    address: "경기 성남시 수정구 여수대로 4",
-    writerNickName:"Edwin",
-    imglink:"https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619"
-  },
-  {
-    num:5,
-    exhibitionTitle: "전시회 제목5",
-    exhibitionDesc: "5555 ⟪대지의 시간⟫은 기후변화와 펜데믹 등 전 지구적 위기 시대를 맞이하여 새로운 시대정신으로 떠오르고 있는 '생태학적 세계관'을 탐색하는 장으로서, '공생', '연결', '균형의 회복'을 지향하는 국내외 작가 16명의 작품과 아카이브를 선보인다.",
-    exhibitionDetaillink : "/exhibition/:id",
-    startDate: "2023-04-12T00:00:00.000Z",
-    endDate: "2023-04-13T00:00:00.000Z",
-    address: "경기 성남시 수정구 여수대로 4",
-    writerNickName:"Edwin",
-    imglink:"https://cdn.mhns.co.kr/news/photo/202109/511451_618343_3128.png"
-  },
-]
