--- src/features/main/mainpageexample/fourthDataList.js
@@ -1,8 +1,0 @@
-export const fourthDataList = [
-  {img : "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200", date:"2023-08-29T00:00:00.000Z", titleKo:"대지의 시간2", titleEn:"The Time of Earth2", location:"과천 : 현대미술관"},
-  {img : "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619", date:"2023-07-01T00:00:00.000Z", titleKo:"대지의 시간3", titleEn:"The Time of Earth3", location:"원주 : 뮤지엄산"},
-  {img : "https://museumnews.kr/wp-content/uploads/2015/06/종이나라박물관-전시-포스터온라인.jpg", date:"2023-12-24T00:00:00.000Z", titleKo:"대지의 시간4", titleEn:"The Time of Earth4", location:"서울 : 예술의전당"},
-  {img : "https://lh3.googleusercontent.com/ci/AEwo86c9tUNWeu-V0oOSi2GEYXzm4sXCmTWVEiTAUzN3NzQVP4oGWi5BJBEIMKgWNLJ1fclBOUqdpm-X=s1200", date:"2023-08-29T00:00:00.000Z", titleKo:"대지의 시간2", titleEn:"The Time of Earth2", location:"과천 : 현대미술관"},
-  {img : "https://culture.seoul.go.kr/cmmn/file/imageSrc.do?fileStreCours=35367259ca6485b8ea26e64a6b235a5388f602a39e9ec217640b29507de1c5f7&streFileNm=a75f86a05a9f5928192d72c7494fb1dfb18a99d5557d38731a0fdd08d3cbb619", date:"2023-07-01T00:00:00.000Z", titleKo:"대지의 시간3", titleEn:"The Time of Earth3", location:"원주 : 뮤지엄산"},
-  {img : "https://museumnews.kr/wp-content/uploads/2015/06/종이나라박물관-전시-포스터온라인.jpg", date:"2023-12-24T00:00:00.000Z", titleKo:"대지의 시간4", titleEn:"The Time of Earth4", location:"서울 : 예술의전당"}
-]
