+++ src/shared/Router.js
@@ -1,5 +1,5 @@
-import React, { useEffect } from "react";
-import { BrowserRouter, Routes, Route, useLocation } from "react-router-dom";
+import React from "react";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { cookies } from "./cookies";
 import Main from "../pages/Main";
 import { ProtectedRoute } from "./protectedRoute";
@@ -13,7 +13,6 @@ import MyPage from "../pages/MyPage";
 import Login from "../pages/Login";
 import Register from "../pages/Register";
 import UpdateExhibition from "../pages/UpdateExhibition";
-import Artgramcopy from "../pages/Artgramcopy";
 
 function Router() {
   const token = cookies.get("access_token");
@@ -27,15 +26,9 @@ function Router() {
       isLogin: false,
     },
     {
-      pathname: "/artgram/origin",
-      element: <Artgramcopy />,
-      isPublic: true,
-      isLogin: false,
-    },
-    {
       pathname: "/artgram/create",
       element: <CreateArtgram />,
-      isPublic: false,
+      isPublic: true,
       isLogin: false,
     },
     {
